+++ catch.h
@@ -45,8 +45,8 @@ class catch_type : public type
 {
 public:
     catch_type() : type("catch", sizeof (struct catcher)) {}
-    unsigned long       mark(ici_obj_t *o) override;
-    void                free(ici_obj_t *o) override;
+    size_t mark(ici_obj_t *o) override;
+    void free(ici_obj_t *o) override;
 };
 
 } // namespace ici
