+++ cfunc.h
@@ -205,10 +205,10 @@ class cfunc_type : public type
 public:
     cfunc_type() : type("func", sizeof (cfunc), type::has_objname | type::has_call) {}
 
-    unsigned long       mark(ici_obj_t *o) override;
-    ici_obj_t *         fetch(ici_obj_t *o, ici_obj_t *k) override;
-    void                objname(ici_obj_t *o, char p[ICI_OBJNAMEZ]) override;
-    int                 call(ici_obj_t *o, ici_obj_t *subject) override;
+    size_t mark(ici_obj_t *o) override;
+    ici_obj_t *fetch(ici_obj_t *o, ici_obj_t *k) override;
+    void objname(ici_obj_t *o, char p[ICI_OBJNAMEZ]) override;
+    int call(ici_obj_t *o, ici_obj_t *subject) override;
 };
 
 } // namespace ici
