+++ file.h
@@ -54,10 +54,10 @@ class file_type : public type
 public:
     file_type() : type("file", sizeof (struct file)) {}
 
-    unsigned long       mark(ici_obj_t *o) override;
-    void                free(ici_obj_t *o) override;
-    int                 cmp(ici_obj_t *o1, ici_obj_t *o2) override;
-    ici_obj_t *         fetch(ici_obj_t *o, ici_obj_t *k) override;
+    size_t mark(ici_obj_t *o) override;
+    void free(ici_obj_t *o) override;
+    int cmp(ici_obj_t *o1, ici_obj_t *o2) override;
+    ici_obj_t *fetch(ici_obj_t *o, ici_obj_t *k) override;
 };
 
 } // namespace ici
