+++ set.h
@@ -30,14 +30,14 @@ class set_type : public type
 public:
     set_type() : type("set", sizeof (struct set), type::has_forall) {}
 
-    unsigned long       mark(ici_obj_t *o) override;
-    void                free(ici_obj_t *o) override;
-    int                 cmp(ici_obj_t *o1, ici_obj_t *o2) override;
-    unsigned long       hash(ici_obj_t *o) override;
-    ici_obj_t *         copy(ici_obj_t *o) override;
-    int                 assign(ici_obj_t *o, ici_obj_t *k, ici_obj_t *v) override;
-    ici_obj_t *         fetch(ici_obj_t *o, ici_obj_t *k) override;
-    int                 forall(ici_obj_t *o) override;
+    size_t mark(ici_obj_t *o) override;
+    void free(ici_obj_t *o) override;
+    int cmp(ici_obj_t *o1, ici_obj_t *o2) override;
+    unsigned long hash(ici_obj_t *o) override;
+    ici_obj_t *copy(ici_obj_t *o) override;
+    int assign(ici_obj_t *o, ici_obj_t *k, ici_obj_t *v) override;
+    ici_obj_t *fetch(ici_obj_t *o, ici_obj_t *k) override;
+    int forall(ici_obj_t *o) override;
 };
 
 } // namespace ici
